Add files via upload
screenshot 1. shows the main menu of the College Event Management Portal after running the Python program. It displays the available options such as viewing events, registering, and checking the participant list.
screenshot 2.shows the list of upcoming college events that the program displays when the user selects the option to view events and given the details of the participant